+++ spec/e2e/RequireTemplate.js
@@ -1,2 +1,2 @@
 var UnitTestSCAD = require('../../src/UnitTestSCAD');
-UnitTestSCAD.run(<config>);
+UnitTestSCAD(<config>);