+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Nothing yet
+### Added
+- Can now invoke `unittestscad` through NodeJS `require`.
 
 ## [2.0.0] - 2017-06-08
 ### Added